# docs/product/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Custara API
  version: 1.0.0-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Custara-Proprietary
  description: |
    REST contract for the Custara V1 customer growth loop.
    Organization context is resolved from the authenticated session/token.
    If the authenticated user belongs to more than one organization, clients must send X-Organization-Id.
    Clients must not send or trust organization IDs in business payloads.
servers:
  - url: https://api.custara.online
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /v1/opportunities/summary:
    get:
      tags: [Opportunities]
      operationId: getOpportunitySummary
      summary: Get aggregate opportunity cards from active customer opportunities
      x-permission: opportunities.read
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: Opportunity summaries
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OpportunitySummary'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/opportunities:
    get:
      tags: [Opportunities]
      operationId: listCustomerOpportunities
      summary: Get the actionable per-customer work queue
      x-permission: opportunities.read
      parameters:
        - $ref: '#/components/parameters/BranchId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/OpportunityType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OpportunityStatus'
        - name: contactable_on_whatsapp
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Opportunity page
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomerOpportunity'
                  meta:
                    $ref: '#/components/schemas/CursorMeta'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/opportunities/{opportunityId}:
    get:
      tags: [Opportunities]
      operationId: getCustomerOpportunity
      summary: Get reason, history, permitted actions, and outcome
      x-permission: opportunities.read
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
      responses:
        '200':
          description: Opportunity detail
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/CustomerOpportunityDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/opportunities/{opportunityId}/actions:
    post:
      tags: [Opportunities]
      operationId: recordGrowthAction
      summary: Record WhatsApp, call, staff follow-up, offer, or campaign action
      x-permission: opportunities.action
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGrowthActionRequest'
      responses:
        '201':
          description: Action recorded and opportunity state updated
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/GrowthAction'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /v1/opportunities/{opportunityId}/dismiss:
    post:
      tags: [Opportunities]
      operationId: dismissCustomerOpportunity
      summary: Dismiss an opportunity with an auditable reason
      x-permission: opportunities.dismiss
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 3
                  maxLength: 255
      responses:
        '200':
          description: Opportunity dismissed
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/CustomerOpportunity'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/opportunity-definitions:
    get:
      tags: [Opportunities]
      operationId: listOpportunityDefinitions
      summary: List configured universal recipes and versions
      x-permission: opportunity_definitions.read
      responses:
        '200':
          description: Definitions
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OpportunityDefinition'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/opportunity-definitions/{definitionId}:
    patch:
      tags: [Opportunities]
      operationId: updateOpportunityDefinition
      summary: Create a new immutable definition version with updated parameters
      x-permission: opportunity_definitions.update
      parameters:
        - name: definitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOpportunityDefinitionRequest'
      responses:
        '200':
          description: New definition version
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/OpportunityDefinition'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BranchId:
      name: X-Branch-Id
      in: header
      required: false
      description: Selected branch. Authorization still validates branch scope.
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    OpportunityId:
      name: opportunityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: Permission or branch scope denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found in the authenticated organization
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State, identity, idempotency, or immutable history conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    CustomerStatus:
      type: string
      enum: [ACTIVE, INACTIVE, MERGED, ANONYMIZED]
...
    OpportunityType:
      type: string
      enum: [INACTIVE, FREQUENCY_DECLINE, CROSS_SELL, NEAR_TIER]
    OpportunityStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, ACTIONED, WON, RESOLVED_ORGANIC, DISMISSED, EXPIRED]
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
        field_errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    CursorMeta:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d{1,2})?$'
          example: '1850000.00'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: IDR
    CustomerSummary:
      type: object
      required: [id, display_name, status, metrics]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        primary_phone:
          type: [string, 'null']
        primary_email:
          type: [string, 'null']
          format: email
        home_branch:
          anyOf:
            - $ref: '#/components/schemas/BranchReference'
            - type: 'null'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        metrics:
          $ref: '#/components/schemas/CustomerMetrics'
        primary_opportunity:
          anyOf:
            - $ref: '#/components/schemas/OpportunityReference'
            - type: 'null'
...
    OpportunitySummary:
      type: object
      required: [definition_id, type, name, priority, customer_count, estimated_value]
      properties:
        definition_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/OpportunityType'
        name:
          type: string
        priority:
          type: integer
        customer_count:
          type: integer
          minimum: 0
        estimated_value:
          $ref: '#/components/schemas/Money'
    CustomerOpportunity:
      type: object
      required: [id, definition_id, type, customer, status, reason_text, reason_data, opened_at, allowed_actions]
      properties:
        id:
          type: string
          format: uuid
        definition_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/OpportunityType'
        customer:
          $ref: '#/components/schemas/CustomerSummary'
        status:
          $ref: '#/components/schemas/OpportunityStatus'
        reason_text:
          type: string
        reason_data:
          type: object
          additionalProperties: true
        estimated_value:
          anyOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        opened_at:
          type: string
          format: date-time
        expires_at:
          type: [string, 'null']
          format: date-time
        allowed_actions:
          type: array
          items:
            type: string
            enum: [OPEN_WHATSAPP, MARK_CONTACTED, CALL, FOLLOW_UP, RECORD_OFFER, DISMISS]
    CustomerOpportunityDetail:
      allOf:
        - $ref: '#/components/schemas/CustomerOpportunity'
        - type: object
          required: [actions, outcomes]
          properties:
            actions:
              type: array
              items:
                $ref: '#/components/schemas/GrowthAction'
            outcomes:
              type: array
              items:
                $ref: '#/components/schemas/GrowthOutcome'
    CreateGrowthActionRequest:
      type: object
      required: [type, channel]
      properties:
        type:
          type: string
          enum: [WHATSAPP_OPENED, WHATSAPP_MARKED_CONTACTED, PHONE_CALL, STAFF_FOLLOW_UP, OFFER_RECORDED, CAMPAIGN_TOUCH]
        channel:
          type: string
          enum: [WHATSAPP, PHONE, IN_PERSON, OTHER]
        branch_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        message_preview:
          type: string
          maxLength: 4000
        performed_at:
          type: string
          format: date-time
    GrowthAction:
      type: object
      required: [id, type, channel, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [WHATSAPP_OPENED, WHATSAPP_MARKED_CONTACTED, PHONE_CALL, STAFF_FOLLOW_UP, OFFER_RECORDED, CAMPAIGN_TOUCH]
        channel:
          type: string
          enum: [WHATSAPP, PHONE, IN_PERSON, OTHER]
        status:
          type: string
          enum: [NOT_STARTED, OPENED, MARKED_CONTACTED, COMPLETED, CANCELLED]
        performed_by_id:
          type: [string, 'null']
          format: uuid
        opened_at:
          type: [string, 'null']
          format: date-time
        performed_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
...
    OpportunityDefinition:
      type: object
      required: [id, key, type, parameters, priority, cooldown_days, attribution_window_days, enabled, version]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        type:
          $ref: '#/components/schemas/OpportunityType'
        parameters:
          $ref: '#/components/schemas/OpportunityParameters'
        priority:
          type: integer
          minimum: 0
        cooldown_days:
          type: integer
          minimum: 0
        attribution_window_days:
          type: integer
          minimum: 1
        enabled:
          type: boolean
        version:
          type: integer
          minimum: 1
    UpdateOpportunityDefinitionRequest:
      type: object
      minProperties: 1
      properties:
        parameters:
          $ref: '#/components/schemas/OpportunityParameters'
        priority:
          type: integer
          minimum: 0
        cooldown_days:
          type: integer
          minimum: 0
        attribution_window_days:
          type: integer
          minimum: 1
        enabled:
          type: boolean
...